Generalize - trait Organism
and restructure a bit.
Still have to fix all the tests and examples
and make a new example for a compound agent
