+++ src/environment.rs
@@ -1,8 +1,8 @@
-use organism::Organism;
+use crate::Genome;
 
 /// A trait that is implemented by user to allow test of the Environment.
-pub trait Environment: Sync {
+pub trait Environment<G: Genome>: Sync {
     /// This test will return the value required by this enviroment to test
     /// against
-    fn test(&self, organism: &mut Organism) -> f64;
+    fn test(&self, organism: &mut G) -> f64;
 }
